Ensure the file is save to current provider selected
Previously, if you opened a file, then went to save-as,
switched to a new provider, and saved (without changing
the name), the file would be saved back to the previous
provider.

This is done on commit instead of every time a tab is
selected so that if you switch tabs and then *don't*
save, when you re-open the dialog your previous provider
will still be the one selected.