Add overwritable flag to file metadata
This is to indicate whether we have a file that we've either
already saved to, or that we open and have permission to
save over.

As for now we don't yet support files without permission to
edit, this is always true on fileChange. This will change
when we support shared files