Add Patch capability
If the CFM requests the savable content from the inner app, and the CFM
is in a state where it could patch (current file is `overwritable` and
provider can `patch`), it tells the inner app through it's `getContent`
request that it will accept a `patchable` result.

If the inner app sends back the content with `isPatch=true`, the CFM
will try to patch.