Add ability for DocServer Provider to open shared content
This will attempt a credential-less open of the specified recordId.
If the record was saved with no owner, it will be opened as usual,
though without a provider or other details in the metadata.

Add client API `openSharedData`. Any file returned from the provider
will be marked with `overwritable: false`.