Envelop all raw data before saving to cloud
This adds a wrapper (an "envelope") around all data before it is passed
to the cloud, which is then unwrapped before it's passed back to the
client. The envelope can contain additional metadata about the file,
such as the appName etc. required by the Doc Server, and, up-next,
details about the shared location and secret shared key.

This means that we are always saving the same content up to all
providers, instead of wrapping content separately for the Doc
Server alone.

When new raw data comes in, either from the innerApp or from the cloud,
we call `createEnvelopedCloudContent`, which will add the additional
metadata (for raw content), or leave the metadata in place (for content
that was already saved in an envelope).