Refactor API to pass only strings to innerApp
This changes the communication between the innerApp and the client so
that only strings are passed back and forth between them, eliminating
any need for the innerApp to know about CloudContent.

Communication of content between client and providers uses only
CloudContent.

By shielding the inner app from CloudContent, we will make it easier
to work with the CloudContent internally (e.g. wrapping data in an
envelope).