If we share, add info to wrapper so we retain it, can re-edit
After sharing, the id of the shared document and the secret shareEditKey
are added to the CloudContent metadata, so that any future save will
include this information in the wrapper.

When we next share (immediately or after loading our personal version of
this file from the cloud) we will have these properties, and can now
update existing shared files on future shares.

Also ensure that we hold onto any preexisting CloudContent when we pull
content from innerApp.