Fix bug caused by content with `prop: undefined`
When we diff two objects together and one has a property with the value
`undefined`, the diff produced is not understood by the Doc Server.

This deletes any undefined keys.

Note that the specific problem will be fixed in the next commit
(`docName` being set to null), but this should fix it for unforeseen
cases.