Save runKey from URL query param, use for saving in LARA
If we launch an activity as a student from LARA, we will have a runKey
in the URL. This runKey is meant to be used when we save back to the
publicly-shared document (which we don't have authorization to). The
document server will then save our changes to a copy of the document, and send back the id of this new document, which we will own.

After we have saved successfully and have the id of the new document,
we can then save via patches, and don't need the runKey.