Don't start patch saving until we have a successful save
When we first launch a document as a student in LARA, we don't have
ownership of the document we have launched. Therefore, we cannot
start using patch saving.

The next commit will make it possible to perform this initial save on
a document we don't own, using the runKey.

I don't see a way to distinguish between a document that we have opened
from the Document Server that we do not have write-access to, and one
that we do, so for now we have to assume that any document we open is
not ours, which means we always have to perform a full-save for the
first save of a session.