Fix renaming doc when opening from read-only prov
Previously, if you opened a document from a read-only provider, changed
the name in the toolbar (or used the `rename` menu option), and then
saved to your own writable-provider, the new file would not use the
new name.

This is now now fixed so that, even though we cannot perform a rename
on a read-only file, we change the `metadata.filename` so that the new
filename will be used on save.

https://www.pivotaltracker.com/story/show/116386817
https://www.pivotaltracker.com/story/show/116342137