Add option to hide menu button
This simply hides the dropdown menu UI, without affecting any other
wrapping or iframing. This allows an iframed app to work as before,
but without visible menu items.

We could already achieve this functionality by setting the menu
to `[]` or `false`, but this is a simpler boolean option that
makes it easier for clients to turn on and off.