Send new `ready` event to client if not opening file
After we `connect` to an app, the app does not know whether the
CFM will next plan on opening a file. Therefore it is hard for the
app to know whether to show a "What would you like to do?"-style
dialog, as CODAP does.

This `ready` event means that there is nothing left that the CFM is
planning to do.

[#100454774]