Call ready after `openedFile` callback; + example
After the CFM has opened a file, when the client calls back without an
error we can call the `ready` event, because there is nothing left to
do. This allows app to use `ready` alone to know the app is ready for
client input (and to hide a splash screen, for instance).

[#100454774]