Correctly set parent folder metadata
If a user named their doc using the title input field, then clicked
Save and selected a folder, the folder selection would not be used.

If the user types in a filename in the Save dialog, the `confirm`
function correctly sets the appropriate folder, but this does not happen
unless the filename has been changed.

The `getStateForFolder` function sets the folder in the state, but this
is never passed into the metadata. This commit fixes that.

[#120377399]