Add ability to load readOnly files from url hash
This gives the readOnly provider the same ability as the other providers,
to load urls of the form `#file=readOnly:some-file`.

Similarly, when read-only files are opened, the url hash is changed to
reflect this, which allows for easier sharing of opened files.

the `_findFile` function recursively searches the file tree to find a
file with the name requested. Filenames are not necessarily unique keys,
but this is up to the maintainer of the read-only file system. Using
names instead of uids means we can keep the existing read-only examples,
and also use human-readable urls (`#file=readOnly:Four%20Seals`).

No attempt is made to worry about loops in the recursive lookup, because
we can't create looped trees.

Currently this won't work if we have multiple read-only providers,
like in the `read-only-with-folders` example page, because we can't
distinguish between them. This will change in the next commit, with
the ability to set the url-display-name of a provider.

[#128470511]