Add ability to open files at arbitrary urls directly
We add a new provider, which is used in a temporary fashion, in that
it is created as needed when the client is asked to open a url
directly, and it lasts only as long as that file's metadata lasts.

The app will automatically try to open urls when it sees the url
fragment

`#file=http...`

The app can also ask the client to open a url directly through the
API, and a CODAP-friendly example of this is added, showing an app
that takes a query parameter and uses it to open a remote file.

[#121532705]