Hash param `newInFolder`: start by editing filename
When we start with the hash param `newInFolder={provider}:{folder}`,
we will want to allow the user to edit the filename, and then
automatically save back to the named folder.

The first step is simply to set the parent folder info on the new
metadata.

The GoogleDrive API will be configured to use this url pattern
when creating new files.

[#129377111]