Fix extensions when downloading files
Previously, we were getting issues where the file "file.json" would be
downloaded as "file.json.json". We now use the `withExtensions`
function to correctly get the filename with the extension.

As part of this, `withExtensions` was made into a class method, and the
other calls to it updated appropriately.

Updated the main example to use an extension (txt), and added an
extension to one of the read-only examples, to show that it wouldn't
be shown in the list, or result in downloading "second-example.txt.txt"

[#125620691]