Allow an array of readable extensions for filtering
Previously, if we set an extension, we would filter the Open list
to just those files with that extension.

This adds an array of alternative extensions, as well as allowing
no extension (must be included explicitly). These will all show up
in the Open list.

Note that when saving a file, the main `extension` property is still
used. We will always save a new file with the single defined extension.

If we open a file with extension "foo", and our set extension
is "bar", we will save "file.foo.bar". (This will be changed in the
next commit.)

[#127512431]