Keep original extension on files we open
If we open "file.foo" from Google Drive, we should save back to
"file.foo", even if our main extension is "bar".

The exception to this is downloading, where we will set the filename
to "file.foo.bar".

The reason for this is that we would *like* to standardize filenames to
our extension (hanse changing when downloading), but we don't want to
change existing filenames.

[#127512431]