Allow array of readable mimeTypes for filtering
Like readableExtensions, this allows an app to define additional
mimeTypes that are readable. This can be useful for supporting
legacy filetypes.

Currently this only applies to GoogleDrive, which is the only
provider that uses mimeTypes when filtering the Open list.

Files will always be saved with the main mimeType, even files that
were originally opened with a different mimeType.

[#115353319]