Save parent id if opened from Google
If we're opening a file directly from Google Drive, we haven't set the
parent info in the metadata, because we haven't been navigating the
list.

Google passes this info down in the file metadata, under an array
called `parents`.

As far as I can see, we can assume that the first element of that array
is always the immediate parent folder.