Client can export files without updating metadata
The new additions to the client API, `saveSecondaryFile` and
`saveSecondaryFileAsDialog` allow the client to save content to any
of the providers (with `saveUnwrapped` capabilities), while not
affecting the existing file metadata.

We also allow the client to define a new extension to be used.

This allows e.g. CODAP to export a separate string of content to a new
.csv file.

[#127512941]