Clone metadata in file view
When we update this view's state, if we haven't changed the filename
then we are operating directly on the client's own metadata.

This was causing a bug where, if we exported a secondary file) which
should create new metadata, but changed the provider without changing
the filename, the client.state.metadata's provider would change too.