Another fix for subfolders and switching providers
This fixes the following three scenarios:

1. Correctly save a file to a subfolder in Drive
2. If you open a read-only file, save it to a subfolder in Drive
3. If you open a file from a subfolder in Drive, you can then open from
another provider without it showing that folder name

This area, as @kswenson has mentioned, has had a lot of churn. The two
main issues are (1) confusion over setting metadata parent and setting
folder state, and (2) all the provider views are initialized at the
same time, not when the user switches between them. This makes sense,
but makes it a little harder to reason correctly about.