Improve Safari Local File save [#135154489]
On Safari, behavior is to automatically download the file with name "Unknown". Due to limitations of Safari, this seems to be the best we can do. To accomplish this we introduce the FileSaver.js library (https://github.com/eligrey/FileSaver.js) which attempts to provide the best cross-browser support possible.

If the 'download' attribute is supported, uses the original code to download the file. If the 'download' attribute is not supported (e.g. Safari), uses the FileSaver.js library to perform the download. WebKit Nightly and the Safari Technology Preview both support the 'download' attribute, so the FileSaver.js library will only be necessary for a limited time.

In conjunction with a change to CODAP to allow import of files without extensions, CODAP will be able to open downloaded "Unknown" files even if they are not manuallly renamed.

Finally, we fix the behavior of the "Save Link As..." context menu item on supported browsers, i.e. not current versions of Safari.