Fix local file tab drop handling
While testing CFM PR #75 (validate document contents) I encountered the fact that files dropped in the drop zone weren't giving the expected error alert message. This turned out to be because the CODAP document drop handler was handling the 'drop' event before the React onDrop() handler. It had been noticed (but not investigated) previously that dropping a file in this dialog sometimes behaved differently than selecting a file from the file chooser and this is presumably the reason for such discrepancies. The fix is to install a DOM event handler on the element rather than using the React onDrop() mechanism.