Fix linked interactives, LARA Embed URL construction
This fixes the bug with linked interactive URLs in the LARA provider using document-server (numeric) ids.

It also fixes a bug with the `getShareLink` for LARA links.

For migrating documents:

The `new_url` format for the migration of authored interactives is this:

https://<CFM_URL>/autolaunch/autolaunch.html
 ?documentId=<url-encoded https://models-resources.concord.org/Flegacy-document-store/<LEGACY_ID>>
 &server=<CODAP_URL> (we had been including cfmBaseUrl, but I don't think we need to)
 &scaling

eg:

```JSON
{
    "id": 207249,
    "old_url": "https://document-store.concord.org/v2/documents/16046/autolaunch?server=https%3A%2F%2Fcodap.concord.org%2Freleases%2Flatest%2F&buttonText=Launch",
    "new_url": "https://cloud-file-manager.concord.org/branch/master/autolaunch/autolaunch.html?documentId=https%3A%2F%2Fmodels-resources.concord.org%2Flegacy-document-store%2F16046&scaling=true&server=https%3A%2F%2Fcodap.concord.org%2Freleases%2Fstaging%2F"
}
```

This commit also allows live-server to bind to 'localhost' for an address.