Changed to always change the Google Drive v2 downfile file location
The v2 api is not returning the correct domain for the download file url making opens fail.  Previously we fell back to changing the domain when the initial domain load failed but a recent change in the behavior of the v2 api causes the error handler not to be called.