Upgrade to Google Drive v3  [#171392351]
Upgraded from v2 to V3 of the Google API.  This upgrade requires a new apiKey option in the config.  This change also added two new strings that needs to be internationalized.

The local secure test port was changed from 8888 to 8080 as that it what is defined in the Google Developer console.